requirements_example.txt
django-compressor==1.1.2